requirements.txt
streamlit~=1.38.0
tensorflow~=2.19.0
Pillow~=10.4.0
numpy~=1.26.2

GitPython~=3.1.43
gitdb~=4.0.11
defusedxml~=0.7.1
ipython~=8.26.0
pip~=24.2
attrs~=23.1.0
distro~=1.8.0
wheel~=0.44.0
PySocks~=1.7.1
protobuf~=5.28.0
cryptography~=42.0.5
filelock~=3.15.4
absl-py~=2.3.0
dill~=0.3.8
gast~=0.6.0
grpcio~=1.72.1
h5py~=3.13.0
pytest~=7.4.4
pytz~=2024.1
rich~=13.7.1
Pygments~=2.18.0
markdown-it-py~=3.0.0
toml~=0.10.2
requests~=2.32.3
keras~=3.10.0
colorama~=0.4.6
PyYAML~=6.0.1
idna~=3.5
multidict~=6.0.4
libclang~=18.1.1
typing_extensions~=4.12.2
smmap~=5.0.1
ml_dtypes~=0.5.1
optree~=0.16.0
packaging~=23.2
psutil~=6.0.0
huggingface-hub~=0.24.6
scipy~=1.14.1
torch~=2.4.1
pandas~=2.2.2
namex~=0.1.0
matplotlib~=3.9.2
opencv-python~=4.10.0.84
mdurl~=0.1.2
charset-normalizer~=3.3.2
setuptools~=75.1.0
cffi~=1.16.0
six~=1.16.0
google-pasta~=0.2.0
regex~=2023.10.3
altair~=5.4.1
jsonschema~=4.23.0
narwhals~=1.6.2
pyarrow~=17.0.0
Jinja2~=3.1.4
referencing~=0.35.1
traitlets~=5.14.3
fsspec~=2024.6.1
config~=0.5.1
aiohttp~=3.9.1
yarl~=1.9.3
MarkupSafe~=2.1.5
SQLAlchemy~=2.0.32
lxml~=5.2.2
numexpr~=2.10.1
python-dateutil~=2.9.0.post0
pydeck~=0.9.1
xxhash~=3.5.0
aiosignal~=1.3.1
frozenlist~=1.4.1
python-multipart~=0.0.9
certifi~=2023.11.17
tzdata~=2024.1
tornado~=6.4.1
h2~=4.1.0
urllib3~=2.1.0
tqdm~=4.67.1
multiprocess~=0.70.16
datasets~=3.6.0
torchvision~=0.19.1
Markdown~=3.8
tenacity~=8.2.3
watchdog~=4.0.2
click~=8.1.7
cachetools~=5.5.0
blinker~=1.8.2
sympy~=1.13.2
langchain~=0.3.5
termcolor~=3.1.0
jsonpointer~=2.4
jsonschema-specifications~=2023.12.1
opt_einsum~=3.4.0
pyparsing~=3.1.4
flatbuffers~=25.2.10
wrapt~=1.16.0
pyreadline3~=3.4.1
astunparse~=1.6.3
tokenizers~=0.20.0
safetensors~=0.4.5
tensorboard~=2.19.0
Werkzeug~=3.1.3
transformers~=4.46.1
beautifulsoup4~=4.12.3
Deprecated~=1.2.14
openai~=1.52.0
torchaudio~=2.4.1
fastapi~=0.104.1
pydantic~=2.9.2
starlette~=0.27.0
uvicorn~=0.30.3
msgpack~=1.1.0